Added character detail

+++ rick-and-morty-ui/src/main/java/com/aqube/ram/ui/characterlist/CharacterListFragment.kt
@@ -48,7 +48,7 @@ class CharacterListFragment : Fragment() {
         characterAdapter.setItemClickListener { character ->
             findNavController().navigate(
                 CharacterListFragmentDirections.actionCharacterListFragmentToCharacterDetailFragment(
-                    character.id
+                    character.id.toLong()
                 )
             )
         }
